chore: ObjC compatibility, unit tests
